src_deploy/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
pandas>=1.5.3
numpy>=1.24.3
jinja2>=3.1.2
pyyaml>=6.0.0
requests>=2.30.0